Add Docker download link to download images without a daemon
This link is a modified version of
https://github.com/moby/moby/blob/master/contrib/download-frozen-image-v2.sh,
which allows for downloading images from a registry over HTTPS, removing the
need for a daemon.

Right now, neither authentication nor other registries besides Docker are
supported.

As input, the link takes an image name (<org>/<name>:<tag>) and outputs the tar
file where the Docker image was stored.