New link: Download container images over HTTP (#4)
* Add Docker download link to download images without a daemon

This link is a modified version of
https://github.com/moby/moby/blob/master/contrib/download-frozen-image-v2.sh,
which allows for downloading images from a registry over HTTPS, removing the
need for a daemon.

Right now, neither authentication nor other registries besides Docker are
supported.

As input, the link takes an image name (<org>/<name>:<tag>) and outputs the tar
file where the Docker image was stored.

* Change download link to take in an output dir instead of tar file

This makes the interface identical to the existing Docker save link. Also, this
is better as a user may download many images to the same directory

* Support authentication and other container registries

* Don't unnecessarily recreate the output tar file

* Use default output directory if none is provided

It technically did use a default output directory before, but this is the cleaner way to do it

* Generalize output parameter

This is required so it integrates nicely with the Nebula modules. The Janus
chain framework requires that parameters that are of the same name have the
same of everything (description, shortcode, default value)

* Move types to a different file and change default value of parameter

* Re-organize functions and constants/regexes

* Refactor to smaller functions for readability

* Remove redirect logic

The Go client does this by default, so it isn't needed

* Format