Add GetLayers and DownloadLayer Docker links
This is a big refactor for the Docker links + types. Most of the logic from the
DockerDownload link has now been separated into a separate Docker Registry
Client. The DockerDownload link is now a minimal version that works exactly
like it did prior to the refactor, now with most registry stuff abstracted.

The GetLayers and DownloadLayer links can now also take advantage of the
registry client to do specific functions instead of needing to download the
entire image.

One optimization that could be done is to pass along the auth token, as right
now a new one is fetched for every layer downloaded.