Add DockerLayerToNP link to convert a Docker layer to NP inputs
This refactors a lot of the ToNPInputs() conversion that's in docker/image.go.
It breaks up functions so they can be used both to collect NP inputs in an
array and via a link to send them right away. The link is concise as it mostly
uses the exact same functionality as is present in the ToNPInputs() function
already.