Add layer downloading for Docker images (#9)
* Add GetLayers and DownloadLayer Docker links

This is a big refactor for the Docker links + types. Most of the logic from the
DockerDownload link has now been separated into a separate Docker Registry
Client. The DockerDownload link is now a minimal version that works exactly
like it did prior to the refactor, now with most registry stuff abstracted.

The GetLayers and DownloadLayer links can now also take advantage of the
registry client to do specific functions instead of needing to download the
entire image.

One optimization that could be done is to pass along the auth token, as right
now a new one is fetched for every layer downloaded.

* Remove fake ID creation when saving layers to a file

This was actually not what docker save does, so let's follow exactly the same
format as docker save here and use the layer hash as the file name. A bonus is
that the code becomes simpler too.

* Add DockerLayerToNP link to convert a Docker layer to NP inputs

This refactors a lot of the ToNPInputs() conversion that's in docker/image.go.
It breaks up functions so they can be used both to collect NP inputs in an
array and via a link to send them right away. The link is concise as it mostly
uses the exact same functionality as is present in the ToNPInputs() function
already.

* Add --max-file-size parameter to limit scanning very large files

Large files are a main reason why this capability runs into memory issues.
Large files are usually binaries that don't have any secrets, so it is
practically useless to scan them anyways. This also helps reduce false
positives.

* Clean up registry_test.go

* Address refactor review comments

* Separate setting auth header to a separate function

Duplicated code and was missed in an area, so a separate function makes this more maintainable

* Log files skipped with Info level

As discussed on Slack, this will help engineers more easily determine if they
should look into scanning these files (e.g. it is an unrecognizable binary)